requirements.txt
asgiref==3.8.1
bleach==6.1.0
Django==4.2.13
django-allauth==0.63.3
django-appconf==1.0.6
django-ckeditor==6.7.1
django-debug-toolbar==4.4.5
django-js-asset==2.2.0
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-robots==6.1
django-simple-captcha==0.6.0
django-taggit==5.0.1
django_compressor==4.5
humanize==4.10.0
pillow==10.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
rcssmin==1.1.2
rjsmin==1.2.2
six==1.16.0
sqlparse==0.5.0
sweetify==2.3.1
tzdata==2024.1
webencodings==0.5.1
